+++ models/contacts.js
@@ -1,12 +1,31 @@
-// const fs = require('fs/promises')
+const fs = require("fs/promises");
+const path = require("path");
+// const { nanoid } = require("nanoid");
 
-const listContacts = async () => {};
+const contactsPath = path.resolve("./models/contacts.json");
 
-const getContactById = async (contactId) => {};
+const listContacts = async () => {
+  const data = await fs.readFile(contactsPath, "utf-8");
+  const contacts = JSON.parse(data);
+  return contacts;
+};
 
-const removeContact = async (contactId) => {};
+const getContactById = async (contactId) => {
+  const data = await fs.readFile(contactsPath, "utf-8");
+  const contacts = JSON.parse(data);
+  const contact = contacts.find((contact) => contact.id === contactId);
+  return contact;
+};
 
 const addContact = async (body) => {};
+
+const removeContact = async (contactId) => {
+  const data = await fs.readFile(contactsPath, "utf-8");
+  const contacts = JSON.parse(data);
+  const newContacts = contacts.filter((contact) => contact.id !== contactId);
+  await fs.writeFile(contactsPath, JSON.stringify(newContacts, null, 2)); // NOTE: WTF?
+  return contacts;
+};
 
 const updateContact = async (contactId, body) => {};
